Functionality for getting and storing
Added code that gets the JSON data of the inputted link and stores them in a SQL database. Also added 2 functions that converts 'str' variable to VARCHAR(1000) to meet the SQL syntax requirement, and a function that returns quotations around the argument if it is a string, verified using regular expressions. 2019-06-16